requirements.txt
PyQt6>=6.6.0
APScheduler>=3.10.0,<4.0.0
python-dotenv>=1.0.0
